shared code bewteen exemptions and marine-licence and moved constants
