+++ src/server/common/constants/coordinate-systems.js
@@ -1,0 +1,24 @@
+export const COORDINATE_SYSTEMS = {
+  WGS84: 'wgs84',
+  OSGB36: 'osgb36'
+}
+
+export const POLYGON_MIN_COORDINATE_POINTS = 3
+
+export const WGS84_CONSTANTS = {
+  MIN_LATITUDE: -90,
+  MAX_LATITUDE: 90,
+  MIN_LONGITUDE: -180,
+  MAX_LONGITUDE: 180,
+  DECIMAL_PLACES: 6
+}
+
+export const OSGB36_CONSTANTS = {
+  MIN_EASTINGS: 0,
+  MAX_EASTINGS: 999999,
+  MIN_NORTHINGS: 0,
+  MAX_NORTHINGS: 9999999,
+  VALID_EASTINGS_LENGTH: 6,
+  MIN_VALID_NORTHINGS_LENGTH: 6,
+  MAX_VALID_NORTHINGS_LENGTH: 7
+}
