+++ src/server/common/validation/centre-coordinates/constants.js
@@ -1,0 +1,33 @@
+import { COORDINATE_SYSTEMS } from '#src/server/common/constants/coordinate-systems.js'
+
+export const centreCoordinatesSettings = {
+  pageTitle: 'Enter the coordinates at the centre point of the site',
+  heading: 'Enter the coordinates at the centre point of the site'
+}
+
+export const centreCoordinatesErrorMessages = {
+  [COORDINATE_SYSTEMS.WGS84]: {
+    LATITUDE_REQUIRED: 'Enter the latitude',
+    LATITUDE_LENGTH: 'Latitude must be between -90 and 90',
+    LATITUDE_NON_NUMERIC: 'Latitude must be a number',
+    LATITUDE_DECIMAL_PLACES:
+      'Latitude must include 6 decimal places, like 55.019889',
+    LONGITUDE_REQUIRED: 'Enter the longitude',
+    LONGITUDE_LENGTH: 'Longitude must be between -180 and 180',
+    LONGITUDE_NON_NUMERIC: 'Longitude must be a number',
+    LONGITUDE_DECIMAL_PLACES:
+      'Longitude must include 6 decimal places, like -1.399500'
+  },
+  [COORDINATE_SYSTEMS.OSGB36]: {
+    EASTINGS_REQUIRED: 'Enter the eastings',
+    EASTINGS_NON_NUMERIC: 'Eastings must be a number',
+    EASTINGS_LENGTH: 'Eastings must be 6 digits',
+    EASTINGS_POSITIVE_NUMBER:
+      'Eastings must be a positive 6-digit number, like 123456',
+    NORTHINGS_REQUIRED: 'Enter the northings',
+    NORTHINGS_NON_NUMERIC: 'Northings must be a number',
+    NORTHINGS_LENGTH: 'Northings must be 6 or 7 digits',
+    NORTHINGS_POSITIVE_NUMBER:
+      'Northings must be a positive 6 or 7-digit number, like 123456'
+  }
+}
