+++ src/server/common/validation/centre-coordinates/validate.js
@@ -1,4 +1,4 @@
-import { COORDINATE_SYSTEMS } from '#src/server/common/constants/exemptions.js'
+import { COORDINATE_SYSTEMS } from '#src/server/common/constants/coordinate-systems.js'
 import { wgs84ValidationSchema } from '#src/server/common/schemas/wgs84.js'
 import { osgb36ValidationSchema } from '#src/server/common/schemas/osgb36.js'
