+++ src/server/common/validation/centre-coordinates/validation-error-cases-osgb36.json
