+++ src/server/common/validation/centre-coordinates/validation-error-cases-wgs84.json
