+++ src/server/common/validation/centre-coordinates/validation-success-cases-osgb36.json
