+++ src/server/common/validation/centre-coordinates/validation-success-cases-wgs84.json
