--- src/server/marine-licence/site-details/centre-coordinates/osgb36.njk
@@ -1,37 +1,0 @@
-{% extends "layouts/coordinate-template.njk" %}
-
-{% set detailsSummary %}
-Help with eastings and northings formats
-{% endset %}
-
-{% set detailsHTML %}
-    {% include "partials/coordinate-entry/osgb36-hint-details.njk" %}
-{% endset %}
-
-{% set introText %}
-  {% include "partials/coordinate-entry/osgb36-intro.njk" %}
-{% endset %}
-
-{% block coordinateFields %}
-    {{ govukInput({
-        label: {
-            text: "Eastings"
-        },
-        id: "eastings",
-        name: "eastings",
-        value: payload.eastings,
-        errorMessage: errors.eastings,
-        classes: "govuk-!-width-one-half"
-    }) }}
-
-    {{ govukInput({
-        label: {
-            text: "Northings"
-        },
-        id: "northings",
-        name: "northings",
-        value: payload.northings,
-        errorMessage: errors.northings,
-        classes: "govuk-!-width-one-half"
-    }) }}
-{% endblock %}
