--- src/server/marine-licence/site-details/centre-coordinates/validation-error-cases-wgs84.json
@@ -1,50 +1,0 @@
-[
-  {
-    "expectedError": "LATITUDE_REQUIRED",
-    "payload": { "longitude": "-0.127800" }
-  },
-  {
-    "expectedError": "LONGITUDE_REQUIRED",
-    "payload": { "latitude": "51.507400" }
-  },
-  {
-    "expectedError": "LATITUDE_NON_NUMERIC",
-    "payload": { "latitude": "abc", "longitude": "-0.127800" }
-  },
-  {
-    "expectedError": "LONGITUDE_NON_NUMERIC",
-    "payload": { "latitude": "51.507400", "longitude": "xyz" }
-  },
-  {
-    "expectedError": "LATITUDE_LENGTH",
-    "payload": { "latitude": "91.000000", "longitude": "-0.127800" }
-  },
-  {
-    "expectedError": "LATITUDE_LENGTH",
-    "payload": { "latitude": "-91.000000", "longitude": "-0.127800" }
-  },
-  {
-    "expectedError": "LONGITUDE_LENGTH",
-    "payload": { "latitude": "51.507400", "longitude": "181.000000" }
-  },
-  {
-    "expectedError": "LONGITUDE_LENGTH",
-    "payload": { "latitude": "51.507400", "longitude": "-181.000000" }
-  },
-  {
-    "expectedError": "LATITUDE_DECIMAL_PLACES",
-    "payload": { "latitude": "55.01988", "longitude": "-0.127800" }
-  },
-  {
-    "expectedError": "LATITUDE_DECIMAL_PLACES",
-    "payload": { "latitude": "55.0198899", "longitude": "-0.127800" }
-  },
-  {
-    "expectedError": "LONGITUDE_DECIMAL_PLACES",
-    "payload": { "latitude": "51.507400", "longitude": "-1.39950" }
-  },
-  {
-    "expectedError": "LONGITUDE_DECIMAL_PLACES",
-    "payload": { "latitude": "51.507400", "longitude": "-1.3995000" }
-  }
-]
