--- src/server/marine-licence/site-details/centre-coordinates/wgs84.njk
@@ -1,37 +1,0 @@
-{% extends "layouts/coordinate-template.njk" %}
-
-{% set detailsSummary %}
-Help with latitude and longitude formats
-{% endset %}
-
-{% set detailsHTML %}
-  {% include "partials/coordinate-entry/wgs84-hint-details.njk" %}
-{% endset %}
-
-{% set introText %}
-  {% include "partials/coordinate-entry/wgs84-intro.njk" %}
-{% endset %}
-
-{% block coordinateFields %}
-    {{ govukInput({
-        label: {
-            text: "Latitude"
-        },
-        id: "latitude",
-        name: "latitude",
-        value: payload.latitude,
-        errorMessage: errors.latitude,
-        classes: "govuk-!-width-one-half"
-    }) }}
-
-    {{ govukInput({
-        label: {
-            text: "Longitude"
-        },
-        id: "longitude",
-        name: "longitude",
-        value: payload.longitude,
-        errorMessage: errors.longitude,
-        classes: "govuk-!-width-one-half"
-    }) }}
-{% endblock %}
