+++ src/server/templates/centre-coordinates/wgs84.njk
